backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Azure OpenAI
openai==1.10.0
azure-identity==1.15.0

# Vector Databases
chromadb==0.4.22
azure-search-documents==11.4.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27

# Data validation and settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1

# HTTP client
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
